refactor: split agents/protocol.rs under the 500-line cap
The base-filter additions tipped crates/trusty-code/src/agents/protocol.rs
to 501 SLOC (1 over the 500-line production cap enforced by
scripts/check_line_cap.sh). Extracted its inline #[cfg(test)] mod tests
block into a sibling agents/protocol_tests.rs, mirroring the identical
workstreams/protocol.rs + protocol_tests.rs split already established in
this crate: #[path = "protocol_tests.rs"] mod tests; keeps the module
logically nested exactly as before (still `tests` from every super::/doc
reference — only which file backs it moved), while the new file's
_tests.rs suffix classifies it as a TEST file (1500 SLOC cap) instead of
counting against the production cap.

protocol.rs: 501 -> 201 SLOC. protocol_tests.rs: 299 SLOC (new, test cap).
line-cap: 8 allowlisted, 0 violations — OK.
